Export registration component to another file
